feat: catch error in get by id methods and prettier code
